032_birthday_wisher_smtp/requirements.txt
anthropic~=0.49.0
openai~=1.66.2
pandas~=2.2.3
python-dotenv~=1.0.1
google-auth~=2.38.0
google-auth-httplib2~=0.2.0
google-auth-oauthlib~=1.2.1
google-api-python-client~=2.163.0
protobuf~=6.30.0